+++ src/commands/plan.rs
@@ -829,7 +829,9 @@ impl Plan {
                         HdlNode::new(HdlSymbol::SystemVerilog(f), node),
                     );
                 }
-                SystemVerilogSymbol::Config(_) | SystemVerilogSymbol::Package(_) => {
+                SystemVerilogSymbol::Config(_)
+                | SystemVerilogSymbol::Package(_)
+                | SystemVerilogSymbol::Interface(_) => {
                     component_pairs.insert(
                         LangIdentifier::Verilog(name.unwrap().clone()),
                         LangIdentifier::Vhdl(vhdl_lib.clone()),