+++ src/core/lang/verilog/token/token.rs
@@ -380,9 +380,8 @@ impl VerilogToken {
             match op_buf.len() {
                 0 => match c {
                     // ambiguous characters...read another character (could be a 2 or 3 length operator)
-                    '(' | ')' | '{' | '}' | '*' | '>' | '<' | '&' | '|' | '=' | '!' | '^' | '~' => {
-                        op_buf.push(train.consume().unwrap())
-                    }
+                    '(' | ')' | '{' | '}' | '*' | '>' | '<' | '&' | '|' | '=' | '!' | '^' | '~'
+                    | '.' => op_buf.push(train.consume().unwrap()),
                     // if it was an operator, take the character and increment the location
                     _ => return Self::match_delimiter(&String::from(train.consume().unwrap())),
                 },