+++ tests/examples/sv/if/apb_if.sv
@@ -1,0 +1,8 @@
+interface apb_if (input pclk);
+	logic [31:0]    paddr;
+	logic [31:0]    pwdata;
+	logic [31:0]    prdata;
+	logic           penable;
+	logic           pwrite;
+	logic           psel;
+endinterface