+++ tests/examples/sv/if/dut.sv
@@ -1,0 +1,7 @@
+module dut (myBus busIf);
+  always @ (posedge busIf.clk)
+    if (busIf.enable)
+      busIf.data <= busIf.data+1;
+    else
+      busIf.data <= 0;
+endmodule