+++ tests/examples/sv/if/my_bus.sv
@@ -1,0 +1,10 @@
+interface myBus (input clk);
+  logic [7:0]  data;
+  logic      enable;
+
+  // From TestBench perspective, 'data' is input and 'write' is output
+  modport TB  (input data, clk, output enable);
+
+  // From DUT perspective, 'data' is output and 'enable' is input
+  modport DUT (output data, input enable, clk);
+endinterface